requirements.txt
numpy==1.24.3
pandas==2.0.3
scikit-learn==1.3.0
tensorflow==2.13.0
opencv-python==4.8.0.74
matplotlib==3.7.2
seaborn==0.12.2
pillow==10.0.0
pytest-cov==4.1.0
flake8==6.0.0
pydantic==2.1.1
python-dotenv==1.0.0
requests==2.31.0
pyyaml==6.0.1
joblib==1.3.1
torchmetrics==1.2.0
mlflow==2.9.2
tqdm==4.66.1
pytest==7.4.3
black==23.11.0